+++ src/shinra/infra/push/__init__.py
@@ -1,0 +1,1 @@
+"""L'invio delle notifiche push verso i browser."""